fix: modem states without CP_OFF
In my case, CP_OFF wasn't needed to maintain state sanity with the PEV. This combination seems to work well, also with the 'autocharge' feature that continually loops through multiple Modem Setup phases. It also fires the disconnect event after being disconnected for 3 seconds instead of 60, so the session is cleared (and access is revoked again) fast. YMMV with these changes.